% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url.R
\name{check_url}
\alias{check_url}
\title{Check URL
A helper function to verify user input before fetching the feed.}
\usage{
check_url(website = "ycombinator.com", rss_table = package_rss)
}
\arguments{
\item{website}{a url of a new source in the format "news.ycombinator.com"}

\item{rss_table}{a dataframe with urls and rss feeds in case you
#need to construct your own out of webisites not in the included database.}
}
\description{
Check URL
A helper function to verify user input before fetching the feed.
}
\examples{
check_url(website = "ycombinator.com", rss_table = package_rss)
}
