% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_urls.R
\name{filter_urls}
\alias{filter_urls}
\title{Filter URLS in the provided database.}
\usage{
filter_urls(
  topic = NULL,
  country = NULL,
  language = NULL,
  rss_table = package_rss
)
}
\arguments{
\item{topic}{the topic of the feed see `show_topics()` for more.}

\item{country}{the country of origin of the feed using two capital
letters, for example "US". See `show_countries()` for more.}

\item{language}{the language of the content of the feed using two
lowercase letters, for example "en". See `show_languages()` for more.}

\item{rss_table}{a dataframe with urls and rss feeds in case you
need to construct your own out of websites not in the included database.}
}
\value{
a tibble filtered according to the given parameters
}
\description{
Filter URLS in the provided database.
}
\examples{
filter_urls(topic = "tech", country = "US", language = "en")
}
