% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news.R
\name{get_news}
\alias{get_news}
\title{Get news}
\usage{
get_news(website = "news.ycombinator.com")
}
\arguments{
\item{website}{a url of a new source in the format "news.ycombinator.com"}
}
\value{
a tibble containing the contents of the rss feed
}
\description{
Get news
}
\examples{
datapath <- system.file("example-data/my-dataset.csv", package = "newscatcheR")
example_data <- readr::read_csv(datapath, col_names = c("url", "rss"))
get_news("news.ycombinator.com")
}
