% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscatcheR-package.R
\docType{package}
\name{newscatcheR-package}
\alias{newscatcheR}
\alias{newscatcheR-package}
\title{newscatcheR: Programmatically Collect Normalized News from (Almost) Any Website}
\description{
Programmatically collect normalized news from (almost) any website. An 'R' clone of the \url{https://github.com/kotartemiy/newscatcher} 'Python' module.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/discindo/newscatcheR/}
  \item Report bugs at \url{https://github.com/discindo/newscatcheR/issues/}
}

}
\author{
\strong{Maintainer}: Novica Nakov \email{nnovica@gmail.com}

Other contributors:
\itemize{
  \item Teofil Nakov \email{teofiln@gmail.com} [contributor]
  \item Artem Bugara \email{bugara.artem@gmail.com} [contributor]
  \item Discindo [copyright holder]
}

}
\keyword{internal}
