% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_rss.R
\docType{data}
\name{package_rss}
\alias{package_rss}
\title{RSS table from python package newscatcher}
\format{
A data frame with 4505 rows and 7 variables:
\describe{
  \item{clean_url}{url of news website}
  \item{language}{the language of the website}
  \item{topic_unified}{the topic of the website}
  \item{main}{main}
  \item{clean_country}{clean_country}
  \item{rss_url}{location of feed}
  \item{GlobalRank}{rank of website}
}
}
\source{
\url{https://github.com/kotartemiy/newscatcher}
}
\usage{
package_rss
}
\description{
A dataset containing sample medical data.
}
\keyword{datasets}
