% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_table.R
\docType{data}
\name{rss_table}
\alias{rss_table}
\title{RSS table from python package newscatcher}
\format{
A data frame with 4241 rows and 2 variables:
\describe{
  \item{url}{url of news website}
  \item{rss_endpoint}{location of feed}
}
}
\source{
\url{https://github.com/kotartemiy/newscatcher}
}
\usage{
rss_table
}
\description{
A dataset containing sample medical data.
}
\keyword{datasets}
