% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{show_countries}
\alias{show_countries}
\title{Show countries
Show all countries in the database.}
\usage{
show_countries(rss_table = package_rss)
}
\arguments{
\item{rss_table}{a dataframe with urls and rss feeds in case you
#need to construct your own out of websites not in the included database.
#Be sure to have the same format as the included data. See `R/package_rss.R`
#for details.}
}
\value{
a character vector of available countries
}
\description{
Show countries
Show all countries in the database.
}
