% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{show_topics}
\alias{show_topics}
\title{Show topics
Show all topics in the database.}
\usage{
show_topics(rss_table = package_rss)
}
\arguments{
\item{rss_table}{a dataframe with urls and rss feeds in case you
#need to construct your own out of websites not in the included database.
#Be sure to have the same format as the included data. See `R/package_rss.R`
#for details.}
}
\value{
a character vector of available topics
}
\description{
Show topics
Show all topics in the database.
}
