% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news.R
\name{get_news}
\alias{get_news}
\title{Get news}
\usage{
get_news(website = "news.ycombinator.com")
}
\arguments{
\item{website}{a url of a new source in the format "news.ycombinator.com"}
}
\value{
a tibble containing the contents of the rss feed
}
\description{
Get news
}
\examples{
Sys.sleep(1) # adding a small time delay to avoid simultaneous posts to the API
get_news(website = "news.ycombinator.com")
}
