% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tld_sources.R
\name{tld_sources}
\alias{tld_sources}
\title{tld_sources
A helper function to explore news sources by country (or other TLD)}
\usage{
tld_sources(tld = NULL)
}
\arguments{
\item{tld}{a country code top level domain in the format ".de"}
}
\value{
a tibble containing the contents of the rss feed
}
\description{
tld_sources
A helper function to explore news sources by country (or other TLD)
}
\examples{
tld_sources(tld = "de")
}
